+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Added nation vocabulary
+  [valentinaB]
 - Alphabetical order for universities
   [valentinaB]
 - Initial release.